refactoring importDiskChain
