+++ packages/xo-server/src/xo-mixins/vmware/importDisksfromDatastore.mjs
@@ -1,11 +1,9 @@
 import { Task } from '@xen-orchestra/mixins/Tasks.mjs'
 import { VDI_FORMAT_QCOW2, VDI_FORMAT_VHD, VHD_MAX_SIZE } from '@xen-orchestra/xapi'
-import parseVmdk from '@xen-orchestra/vmware-explorer/parsers/vmdk.mjs'
 import { ReadAhead } from '@xen-orchestra/disk-transform'
 import { toVhdStream } from 'vhd-lib/disk-consumer/index.mjs'
 import { NbdDisk } from '@vates/nbd-client/NbdDisk.mjs'
 import { createLogger } from '@xen-orchestra/log'
-import NbdClient from '@vates/nbd-client'
 import { toQcow2Stream } from '@xen-orchestra/qcow2'
 
 const { warn } = createLogger('xo:importdiskfromdatastore')
@@ -34,69 +32,13 @@ async function importDiskChain({ esxi, sr, vm, chainByNode, userdevice, vmId }) 
         'vddk import does not support importing multiple snapshots. Coalesce the non imported snapshot into one or force a full import'
       )
     }
-    let nbdClient
-    try {
-      // get the changed blocks of the next snapshot
-      const existingDisk = chainByNode[previouslyImportedIndex]
-      existingVdi = diskIsAlreadyImported(existingVdis, existingDisk)
-      Task.info(`found a previous import`, { vdiRef: existingVdi.$ref })
-      const nbdInfoSpawn = await esxi.spanwNbdKitProcess(vmId, `[${datastoreName}] ${diskPath}`, {
-        singleLink: true,
-      })
-      Task.info(`nbd server for data map spawned`)
-      nbdClient = new NbdClient(nbdInfoSpawn.nbdInfos)
 
-      await nbdClient.connect()
-      Task.info(`nbd client for data map connected`)
-      dataMap = await nbdClient.getMap()
-      Task.info(
-        `got the data map of the single disk in ${Math.round((Date.now() - start) / 1000)} seconds ,${dataMap.length} blocks`
-      )
-    } catch (error) {
-      const descriptorResponse = await esxi.download(datastoreName, diskPath, '0-512')
-      const descriptorBlob = await new Response(descriptorResponse.body).blob()
-      const descriptorArrayBuffer = await descriptorBlob.arrayBuffer()
-      const descriptorBytes = new Uint8Array(descriptorArrayBuffer).slice(0, 512)
-      const descriptorText = new TextDecoder('utf-8').decode(descriptorBytes)
+    // get the changed blocks of the next snapshot
+    const existingDisk = chainByNode[previouslyImportedIndex]
+    existingVdi = diskIsAlreadyImported(existingVdis, existingDisk)
+    Task.info(`found a previous import`, { vdiRef: existingVdi.$ref })
 
-      const parsedDescriptor = parseVmdk(descriptorText)
-
-      const diskPathArray = diskPath.split('/')
-      const extentPath = diskPathArray.slice(0, -1).join('/') + '/' + parsedDescriptor.fileName
-
-      const extentHeaderResponse = await esxi.download(datastoreName, extentPath, `0-2048`)
-      const extentHeaderBlob = await new Response(extentHeaderResponse.body).blob()
-      const extentHeaderBuffer = Buffer.from(await extentHeaderBlob.arrayBuffer())
-
-      const extentNumGdEntries = extentHeaderBuffer.readUInt32LE(24)
-
-      const extentGDResponse = await esxi.download(datastoreName, extentPath, `2048-${2048 + extentNumGdEntries * 4}`)
-      const extentGDBlob = await new Response(extentGDResponse.body).blob()
-      const extentGDBuffer = Buffer.from(await extentGDBlob.arrayBuffer())
-
-      dataMap = []
-      let offset = 0
-      for (let i = 0; i < extentNumGdEntries; i++) {
-        const extentGDE = extentGDBuffer.readUInt32LE(i * 4)
-        if (extentGDE !== 0) {
-          dataMap.push({
-            offset: offset,
-            length: 4096 * 512,
-            type: 0,
-          })
-        }
-
-        offset += 4096 * 512
-      }
-
-      Task.warning('error while getting the map of a snapshot, fall back to a full import', error)
-      // throw error
-    } finally {
-      await nbdClient.disconnect()
-      await esxi
-        .killNbdServer(vmId, `[${datastoreName}] ${diskPath}`, { singleLink: true })
-        .catch(err => Task.warning('error while stopping nbdkit server for the snapshot', err))
-    }
+    dataMap = await esxi.getDataMap(vmId, datastoreName, diskPath)
   } else {
     Task.info(`no reference disk found, fall back a full import`)
   }
